Rename useDebounce hook and improve its tests
